Finalizados métodos get e post para comunicação com API

+++ src/main/java/com/lucasgiavaroti/spring_cardapio/entity/FoodRequestDTO.java
@@ -1,4 +1,4 @@
 package com.lucasgiavaroti.spring_cardapio.entity;
 
-public record FoodRequestDTO(String title, String image, int price) {
+public record FoodRequestDTO(String title, String image, Double price) {
 }
