+++ src/main/java/com/lucasgiavaroti/spring_cardapio/entity/FoodResponseDTO.java
@@ -1,6 +1,6 @@
 package com.lucasgiavaroti.spring_cardapio.entity;
 
-public record FoodResponseDTO(Long id, String title, String image, Integer price) {
+public record FoodResponseDTO(Long id, String title, String image, Double price) {
     public FoodResponseDTO(Food food){
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
